// packages/desktop/src/renderer/components/Editor/MonacoEditor.d.ts
interface MonacoEditorProps {
    value: string;
    onChange: (value: string) => void;
    language?: string;
    theme?: string;
    readOnly?: boolean;
}
export declare const MonacoEditor: ({ value, onChange, language, theme, readOnly, }: MonacoEditorProps) => import("react/jsx-runtime").JSX.Element;
export {};
//# sourceMappingURL=MonacoEditor.d.ts.map